refactor: move `InvalidCastException`
