+++ src/Illuminate/Contracts/Validation/Validator.php
@@ -38,7 +38,7 @@ interface Validator extends MessageProvider
      * @return array<string, mixed>
      *
      * @throws \Illuminate\Validation\ValidationException
-     * @throws \Illuminate\Validation\InvalidCastException
+     * @throws \Illuminate\Support\Exceptions\InvalidCastException
      */
     public function validateAndCast();
 
@@ -50,7 +50,7 @@ interface Validator extends MessageProvider
      * @return mixed
      *
      * @throws \Illuminate\Validation\ValidationException
-     * @throws \Illuminate\Validation\InvalidCastException
+     * @throws \Illuminate\Support\Exceptions\InvalidCastException
      */
     public function casted($key = null, $default = null);
